+++ src/PhpProject/Reader/GanttProject.php
@@ -18,6 +18,9 @@
 namespace PhpOffice\PhpProject\Reader;
 
 use PhpOffice\PhpProject\PhpProject;
+use PhpOffice\PhpProject\Resource;
+use PhpOffice\PhpProject\Task;
+use PhpOffice\PhpProject\Shared\XMLReader;
 
 /**
  * GanttProject
@@ -38,19 +41,18 @@ class GanttProject
     /**
      * Create a new GanttProject
      */
-    public function __construct()
+    public function __construct ()
     {
-        $this->phpProject    = new PhpProject();
+        $this->phpProject = new PhpProject();
     }
     /**
      *
-     * @todo
      * @param string $pFilename
      * @return PHPProject
      */
-    public function canRead($pFilename)
+    public function canRead ($pFilename)
     {
-        if (file_exists($pFilename)) {
+        if (file_exists($pFilename) && is_readable($pFilename)) {
             return true;
         }
         return false;
@@ -58,14 +60,154 @@ class GanttProject
     
     /**
      * 
-     * @todo
      * @param string $pFilename
+     * @throws \Exception
      * @return PHPProject
      */
-    public function load($pFilename)
+    public function load ($pFilename)
     {
-        if (file_exists($pFilename)) {
-            return $this->phpProject;
+        if (!file_exists($pFilename) || !is_readable($pFilename)) {
+            throw new \Exception('The file is not accessible.');
+        }
+        $content = file_get_contents($pFilename);
+        $oXML = new XMLReader();
+        $oXML->getDomFromString($content);
+        
+        $readMethods = array(
+            'description' => 'readNodeDescription',
+            'tasks' => 'readNodeTasks',
+            'resources' => 'readNodeResources',
+            'allocations' => 'readNodeAllocations',
+        );
+        
+        $oNodes = $oXML->getElements('*');
+        if ($oNodes->length > 0) {
+            foreach ($oNodes as $oNode) {
+                if (array_key_exists($oNode->nodeName, $readMethods)) {
+                    $readMethod = $readMethods[$oNode->nodeName];
+                    $this->$readMethod($oXML, $oNode);
+                }
+            }
+        }
+        
+        return $this->phpProject;
+    }
+    
+    /**
+     * Node "Description"
+     * @param XMLReader $oXML
+     * @param \DOMElement $domNode
+     */
+    private function readNodeDescription (XMLReader $oXML, \DOMElement $domNode)
+    {
+        $this->phpProject->getProperties()->setDescription($domNode->nodeValue);
+    }
+    
+    /**
+     * Node "Tasks"
+     * @param XMLReader $oXML
+     * @param \DOMElement $domNode
+     */
+    private function readNodeTasks (XMLReader $oXML, \DOMElement $domNode)
+    {
+        $oNodes = $oXML->getElements('*', $domNode);
+        if ($oNodes->length > 0) {
+            foreach ($oNodes as $oNode) {
+                if ($oNode->nodeName == 'task') {
+                    $oTask = $this->phpProject->createTask();
+                    $this->readNodeTask($oXML, $oNode, $oTask);
+                }
+            }
+        }
+    }
+    
+    /**
+     * Node "Task"
+     * @param XMLReader $oXML
+     * @param \DOMElement $domNode
+     */
+    private function readNodeTask (XMLReader $oXML, \DOMElement $domNode, Task $oTask)
+    {
+        // Attributes
+        $oTask->setIndex($domNode->getAttribute('id'));
+        $oTask->setName($domNode->getAttribute('name'));
+        $oTask->setStartDate($domNode->getAttribute('start'));
+        $oTask->setDuration($domNode->getAttribute('duration'));
+        $oTask->setProgress($domNode->getAttribute('complete'));
+        
+        // SubNodes
+        $oNodes = $oXML->getElements('*', $domNode);
+        if ($oNodes->length > 0) {
+            foreach ($oNodes as $oNode) {
+                if ($oNode->nodeName == 'task') {
+                    $oTaskChild = $oTask->createTask();
+                    $this->readNodeTask($oXML, $oNode, $oTaskChild);
+                }
+            }
+        }
+    }
+    
+    /**
+     * Node "Resources"
+     * @param XMLReader $oXML
+     * @param \DOMElement $domNode
+     */
+    private function readNodeResources (XMLReader $oXML, \DOMElement $domNode)
+    {
+        $oNodes = $oXML->getElements('*', $domNode);
+        if ($oNodes->length > 0) {
+            foreach ($oNodes as $oNode) {
+                if ($oNode->nodeName == 'resource') {
+                    $oResource = $this->phpProject->createResource();
+                    $this->readNodeResource($oXML, $oNode, $oResource);
+                }
+            }
+        }
+    }
+    /**
+     * Node "Resource"
+     * @param XMLReader $oXML
+     * @param \DOMElement $domNode
+     */
+    private function readNodeResource (XMLReader $oXML, \DOMElement $domNode, Resource $oResource)
+    {
+        // Attributes
+        $oResource->setIndex($domNode->getAttribute('id'));
+        $oResource->setTitle($domNode->getAttribute('name'));
+    }
+    
+    /**
+     * Node "Allocations"
+     * @param XMLReader $oXML
+     * @param \DOMElement $domNode
+     */
+    private function readNodeAllocations (XMLReader $oXML, \DOMElement $domNode)
+    {
+        $oNodes = $oXML->getElements('*', $domNode);
+        if ($oNodes->length > 0) {
+            foreach ($oNodes as $oNode) {
+                if ($oNode->nodeName == 'allocation') {
+                    $this->readNodeAllocation($oXML, $oNode);
+                }
+            }
+        }
+    }
+    /**
+     * Node "Allocation"
+     * @param XMLReader $oXML
+     * @param \DOMElement $domNode
+     */
+    private function readNodeAllocation (XMLReader $oXML, \DOMElement $domNode)
+    {
+        // Attributes
+        $idTask = $domNode->getAttribute('task-id');
+        $idResource = $domNode->getAttribute('resource-id');
+        
+        $oResource = $this->phpProject->getResourceFromIndex($idResource);
+        $oTask = $this->phpProject->getTaskFromIndex($idTask);
+        
+        if ($oResource instanceof Resource && $oTask instanceof Task) {
+            $oTask->addResource($oResource);
         }
     }
 }