+++ tests/PhpProject/Tests/PhpProjectTest.php
@@ -77,6 +77,18 @@ class PhpProjectTest extends \PHPUnit_Framework_TestCase
         $this->assertInstanceOf('PhpOffice\\PhpProject\\Resource', $object->getResource(0));
     }
     
+    public function testResourceFromIndex()
+    {
+    	$object = new PhpProject();
+    	$oResource1 = $object->createResource();
+    	$oResource1->setIndex(2);
+    	$oResource2 = $object->createResource();
+    	$oResource2->setIndex(4);
+    	 
+    	$this->assertInstanceOf('PhpOffice\\PhpProject\\Resource', $object->getResourceFromIndex(2));
+    	$this->assertNull($object->getResourceFromIndex(1));
+    }
+    
     /**
      * @expectedException \Exception
      * @expectedExceptionMessage Resource index is out of bounds.
@@ -127,6 +139,20 @@ class PhpProjectTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(0, $object->getActiveTaskIndex());
         $this->assertInternalType('array', $object->getAllTasks());
         $this->assertInstanceOf('PhpOffice\\PhpProject\\Task', $object->getActiveTask());
+    }
+    
+    public function testTaskFromIndex()
+    {
+    	$object = new PhpProject();
+    	$oTask1 = $object->createTask();
+    	$oTask1->setIndex(2);
+    	$oTask11 = $oTask1->createTask();
+    	$oTask11->setIndex(22);
+    	$oTask2 = $object->createTask();
+    	$oTask2->setIndex(4);
+    	
+    	$this->assertInstanceOf('PhpOffice\\PhpProject\\Task', $object->getTaskFromIndex(2));
+    	$this->assertNull($object->getTaskFromIndex(1));
     }
     
     /**