+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
-## 0.1.0 - Not Released
+## 0.1.0 - 2014-08-08
 
 ### Features
-- GanttProject Writer - @Progi1984 GH-1
 - Support of Composer - @Progi1984 GH-7 GH-9
 - Support of namespaces - @Progi1984 GH-12
+- GanttProject Writer - @Progi1984 GH-1
+- GanttProject Reader - @Progi1984 GH-3
 
 ### Bugfix